adding support for serializable closure 2.0 + housekeeping

+++ tests/HelperTest.php
@@ -4,11 +4,12 @@ namespace Jobby\Tests;
 
 use Jobby\Helper;
 use Jobby\Jobby;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @coversDefaultClass Jobby\Helper
  */
-class HelperTest extends \PHPUnit_Framework_TestCase
+class HelperTest extends TestCase
 {
     /**
      * @var Helper
@@ -33,8 +34,10 @@ class HelperTest extends \PHPUnit_Framework_TestCase
     /**
      * {@inheritdoc}
      */
-    protected function setUp()
+    protected function setUp(): void
     {
+        parent::setUp();
+
         $this->helper = new Helper();
         $this->tmpDir = $this->helper->getTempDir();
         $this->lockFile = $this->tmpDir . '/test.lock';
@@ -44,9 +47,11 @@ class HelperTest extends \PHPUnit_Framework_TestCase
     /**
      * {@inheritdoc}
      */
-    protected function tearDown()
+    protected function tearDown(): void
     {
         unset($_SERVER['APPLICATION_ENV']);
+
+        parent::tearDown();
     }
 
     /**
@@ -133,7 +138,10 @@ class HelperTest extends \PHPUnit_Framework_TestCase
      */
     public function testLockLifetimeShouldBeZeroIfFileDoesNotExists()
     {
-        unlink($this->lockFile);
+        if (file_exists($this->lockFile)) {
+            unlink($this->lockFile);
+        }
+
         $this->assertFalse(file_exists($this->lockFile));
         $this->assertEquals(0, $this->helper->getLockLifetime($this->lockFile));
     }
@@ -182,19 +190,22 @@ class HelperTest extends \PHPUnit_Framework_TestCase
 
     /**
      * @covers ::releaseLock
-     * @expectedException \Jobby\Exception
      */
     public function testReleaseNonExistin()
     {
+        $this->expectException(\Jobby\Exception::class);
+
         $this->helper->releaseLock($this->lockFile);
     }
 
     /**
      * @covers ::acquireLock
-     * @expectedException \Jobby\InfoException
      */
     public function testExceptionIfAquireFails()
     {
+        $this->expectException(\Jobby\InfoException::class);
+
+        touch($this->lockFile);
         $fh = fopen($this->lockFile, 'r+');
         $this->assertTrue(is_resource($fh));
 
@@ -206,10 +217,11 @@ class HelperTest extends \PHPUnit_Framework_TestCase
 
     /**
      * @covers ::acquireLock
-     * @expectedException \Jobby\Exception
      */
     public function testAquireLockShouldFailOnSecondTry()
     {
+        $this->expectException(\Jobby\Exception::class);
+
         $this->helper->acquireLock($this->lockFile);
         $this->helper->acquireLock($this->lockFile);
     }
@@ -280,14 +292,14 @@ class HelperTest extends \PHPUnit_Framework_TestCase
 
         $host = $helper->getHost();
         $email = "jobby@$host";
-        $this->assertContains('job', $mail->getSubject());
-        $this->assertContains("[$host]", $mail->getSubject());
+        $this->assertStringContainsString('job', $mail->getSubject());
+        $this->assertStringContainsString("[$host]", $mail->getSubject());
         $this->assertEquals(1, count($mail->getFrom()));
         $this->assertEquals('jobby', current($mail->getFrom()));
         $this->assertEquals($email, current(array_keys($mail->getFrom())));
         $this->assertEquals($email, current(array_keys($mail->getSender())));
-        $this->assertContains($config['output'], $mail->getBody());
-        $this->assertContains('message', $mail->getBody());
+        $this->assertStringContainsString($config['output'], $mail->getBody());
+        $this->assertStringContainsString('message', $mail->getBody());
     }
 
     /**
@@ -297,7 +309,10 @@ class HelperTest extends \PHPUnit_Framework_TestCase
     {
         $nullTransport = new \Swift_NullTransport();
 
-        return $this->getMock('Swift_Mailer', [], [$nullTransport]);
+        return $this->getMockBuilder(\Swift_Mailer::class)
+            ->setConstructorArgs([$nullTransport])
+            ->onlyMethods(['send'])
+            ->getMock();
     }
 
     /**
@@ -306,7 +321,9 @@ class HelperTest extends \PHPUnit_Framework_TestCase
     public function testItReturnsTheCorrectNullSystemDeviceForUnix()
     {
         /** @var Helper $helper */
-        $helper = $this->getMock("\\Jobby\\Helper", ["getPlatform"]);
+        $helper = $this->getMockBuilder(Helper::class)
+            ->onlyMethods(['getPlatform'])
+            ->getMock();
         $helper->expects($this->once())
             ->method("getPlatform")
             ->willReturn(Helper::UNIX);
@@ -320,7 +337,9 @@ class HelperTest extends \PHPUnit_Framework_TestCase
     public function testItReturnsTheCorrectNullSystemDeviceForWindows()
     {
         /** @var Helper $helper */
-        $helper = $this->getMock("\\Jobby\\Helper", ["getPlatform"]);
+        $helper = $this->getMockBuilder(Helper::class)
+            ->onlyMethods(['getPlatform'])
+            ->getMock();
         $helper->expects($this->once())
                ->method("getPlatform")
                ->willReturn(Helper::WINDOWS);
